+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add support for replication mode, [PR-53](https://github.com/reductstore/reduct-rs/pull/53)
+- Add `ResourceStatus` enum for non-blocking deletion support (READY/DELETING), [PR-XX](https://github.com/reductstore/reduct-rs/pull/XX)
 
 ## 1.17.2 - 2025-12-17